Backend/src/main/java/com/dh/dhbooking/dto/CategoryDTO.java
package com.dh.dhbooking.dto;



public class CategoryDTO {

    private Integer id;
    private String title ;
    private String description;
    private String imageUrl;

    private Boolean status;


    public CategoryDTO() {
    }

    public CategoryDTO(String title, String description, String imageUrl) {
        this.title = title;
        this.description = description;
        this.imageUrl = imageUrl;
    }

    public Integer getId() {
        return id;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public String getImageUrl() {
        return imageUrl;
    }

    public void setImageUrl(String imageUrl) {
        this.imageUrl = imageUrl;
    }

    public Boolean getStatus() {
        return status;
    }

    public void setStatus(Boolean status) {
        this.status = status;
    }
}
